separate the client hints
