+++ src/ua_generator/__init__.py
@@ -3,8 +3,8 @@ Random User-Agent
 Copyright: 2022 Ekin Karadeniz (github.com/iamdual)
 License: Apache License 2.0
 """
-from . import useragent
+from . import user_agent
 
 
 def generate(**kwargs):
-    return useragent.UserAgent().generate(**kwargs)
+    return user_agent.UserAgent().generate(**kwargs)
